Refactor services to interfaces
