+++ oriedita/src/main/java/oriedita/editor/factory/HistoryStateFactory.java
@@ -3,6 +3,7 @@ package oriedita.editor.factory;
 import dagger.Module;
 import dagger.Provides;
 import oriedita.editor.service.HistoryState;
+import oriedita.editor.service.impl.DequeHistoryState;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
@@ -11,11 +12,11 @@ import javax.inject.Singleton;
 public class HistoryStateFactory {
     @Provides @Singleton @Named("normal")
     public static HistoryState normalHistoryState() {
-        return new HistoryState();
+        return new DequeHistoryState();
     }
 
     @Provides @Singleton @Named("aux")
     public static HistoryState auxHistoryState() {
-        return new HistoryState();
+        return new DequeHistoryState();
     }
 }
