+++ oriedita/src/main/java/oriedita/editor/task/CheckCAMVTask.java
@@ -1,16 +1,8 @@
 package oriedita.editor.task;
 
 import org.tinylog.Logger;
-import oriedita.editor.canvas.CreasePattern_Worker;
 import oriedita.editor.databinding.CanvasModel;
-import oriedita.editor.service.SingleTaskExecutorService;
 import origami.crease_pattern.FoldLineSet;
-
-import javax.inject.Named;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 
 public class CheckCAMVTask implements OrieditaTask {
     private final FoldLineSet foldLineSet;
