+++ oriedita/src/test/java/oriedita/editor/save/SaveTest.java
@@ -11,9 +11,10 @@ import oriedita.editor.databinding.*;
 import oriedita.editor.drawing.tools.Camera;
 import oriedita.editor.exception.FileReadingException;
 import oriedita.editor.service.FileSaveService;
-import oriedita.editor.service.HistoryState;
 import oriedita.editor.service.ResetService;
-import oriedita.editor.service.SingleTaskExecutorService;
+import oriedita.editor.service.impl.FileSaveServiceImpl;
+import oriedita.editor.service.impl.DequeHistoryState;
+import oriedita.editor.service.impl.SingleTaskExecutorServiceImpl;
 import oriedita.editor.text.Text;
 import origami.Epsilon;
 import origami.crease_pattern.FoldLineSet;
@@ -47,9 +48,9 @@ public class SaveTest {
         FoldedFigureModel foldedFigureModel = new FoldedFigureModel();
         SelectedTextModel textModel = new SelectedTextModel();
         TextWorker textWorker = new TextWorker();
-        mainCreasePatternWorker = new CreasePattern_Worker(creasePatternCamera, new HistoryState(), new HistoryState(), new FoldLineSet(), new FoldLineSet(), new SingleTaskExecutorService(), canvasModel, applicationModel, gridModel, foldedFigureModel, fileModel, null, null, textWorker, textModel);
+        mainCreasePatternWorker = new CreasePattern_Worker(creasePatternCamera, new DequeHistoryState(), new DequeHistoryState(), new FoldLineSet(), new FoldLineSet(), new SingleTaskExecutorServiceImpl(), canvasModel, applicationModel, gridModel, foldedFigureModel, fileModel, null, null, textWorker, textModel);
         ResetService resetService = () -> {};
-        fileSaveService = new FileSaveService(null, creasePatternCamera, mainCreasePatternWorker, null, fileModel, applicationModel, canvasModel, new FoldedFiguresList(), resetService, null);
+        fileSaveService = new FileSaveServiceImpl(null, creasePatternCamera, mainCreasePatternWorker, null, fileModel, applicationModel, canvasModel, new FoldedFiguresList(), resetService, null);
     }
 
     @ParameterizedTest
